Add files via upload
The food donation app is an innovative mobile application that connects 
food donors with food recipients in order to reduce food waste and address
food insecurity in communities. The app provides an efficient, user-friendly
 platform that enables food donors to donate surplus food to local food 
banks, charities, and other organizations that can then distribute the food to
those in need.

The app features an easy-to-use interface that allows donors to    quickly upload information about the type and quantity of food they want to donate.    Donors can also schedule pick-up times with the food recipient .

Food recipients can search for available food donations in their local area and request specific types of food. They can also receive notifications about new donations and pick-up times, ensuring that they are aware of all
available resources.

One of the key benefits of the food donation app is that it helps to reduce food waste, which is a major problem in many communities.By connecting        food donors with food recipients, the app helps to ensure that surplus food
is being used effectively and not wasted.

Additionally, the app helps to address food insecurity by providing a reliable source of food for those in need. By facilitating the distribution of surplus food to food banks and other organizations, the app helps to ensure that
people who are struggling to access food have access to nutritious and healthy options.

Overall, the food donation app is an important tool in the fight against food waste and food insecurity. By connecting food donors with food recipients and providing a platform for easy, efficient, and transparent donations, the app helps to ensure that surplus food is used effectively and that those who  need it most have access to nutritious and healthy options.